Homework 6 updated
The screen now clears if "=" was just pressed and a new number is typed
in, beginning a new mathematical operation.  The Minus function is still
acting very strange, because I'm not able to figure out how
DoMathWithSavedOperator() should work.  I also added messages when the
user divides by zero, but I'm not able to get the text to show up on the
screen.  I made message boxes as stand-ins.